통합에러 해결
