+++ package.json
@@ -1,36 +1,74 @@
 {
 
   "name": "minimart-api",
+
   "version": "1.0.0",
+
   "description": "",
-   "main": "app.js",
+
+  "main": "app.js",
+
   "scripts": {
+
     "start": "nodemon app.js",
+
     "test": "echo \"Error: no test specified\" && exit 1"
+
   },
+
   "keywords": [],
+
   "author": "",
+
   "license": "ISC",
+
   "type": "commonjs",
+
   "dependencies": {
+
     "axios": "^1.11.0",
+
     "bcrypt": "^6.0.0",
+
     "cookie-parser": "^1.4.7",
+
     "cors": "^2.8.5",
+
     "dotenv": "^17.2.1",
+
     "express": "^5.1.0",
+
     "express-session": "^1.18.2",
+
     "jsonwebtoken": "^9.0.2",
+
     "morgan": "^1.10.1",
+
     "multer": "^2.0.2",
+
     "mysql2": "^3.14.3",
+
     "passport": "^0.7.0",
+
     "passport-google-oauth20": "^2.0.0",
+
     "passport-local": "^1.0.0",
+
     "sequelize": "^6.37.7",
-    "socket.io": "^4.8.1"
+
+    "socket.io": "^4.8.1",
+
+    "swagger-jsdoc": "^6.2.8",
+
+    "swagger-ui-express": "^5.0.1"
+
   },
+
   "devDependencies": {
+
     "nodemon": "^3.1.10"
+
   }
+
 }
+
