+++ routes/auth/auth.js
@@ -1,10 +1,11 @@
 const express = require('express')
 const google = require('./auth_google')
+const kakao = require('./auth_kakao')
 
 const router = express.Router()
 
 router.use('/google', google)
-// router.use(카카오)
+router.use('/kakao', kakao)
 // router.use(로컬)
 
 module.exports = router
